requirements.txt
flet==0.27.1
pymongo==4.11.2
python-dotenv